django/requirements.txt
Django==4.2.0
daphne>=4.0.0,<5.0.0
django-environ>=0.9.0
django-cors-headers>=4.2.0
djangorestframework>=3.11,<4.0
channels>=4.0,<5.0
channels-redis>=4.0,<5.0
psycopg2-binary>=2.8
redis>=4.0
requests>=2.31.0
celery>=5.4
